Move localmutexes.go to GCE-specific cloudup
