+++ upup/pkg/fi/cloudup/gce/gce_cloud.go
@@ -36,7 +36,6 @@ import (
 	"k8s.io/kops/dnsprovider/pkg/dnsprovider"
 	"k8s.io/kops/dnsprovider/pkg/dnsprovider/providers/google/clouddns"
 	"k8s.io/kops/pkg/apis/kops"
-	"k8s.io/kops/pkg/mutexes"
 	"k8s.io/kops/upup/pkg/fi"
 	"k8s.io/kops/upup/pkg/fi/cloudup/gce/gcemetadata"
 )
@@ -60,8 +59,8 @@ type GCECloud interface {
 }
 
 // MutexForProjectIAM returns a mutex to prevent local concurrent operations on project IAM.
-func MutexForProjectIAM(projectID string) mutexes.LocalMutex {
-	return mutexes.InProcess.Get("iam/projects/" + projectID)
+func MutexForProjectIAM(projectID string) LocalMutex {
+	return InProcessMutex.Get("iam/projects/" + projectID)
 }
 
 type gceCloudImplementation struct {
